test: migrate LocationAutocomplete, LocationAutocompleteWithGoogleLoader tests
